+++ src/GUI/Converters/UriToBitmapImageConverter.cs
@@ -7,6 +7,7 @@ namespace DivinityModManager.Converters;
 internal class UriToBitmapImageConverter : IValueConverter
 {
 	private static readonly Dictionary<string, BitmapImage> ImageCache = new(StringComparer.OrdinalIgnoreCase);
+	private static readonly object ImageCacheLock = new();
 
 	public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 	{
@@ -15,18 +16,35 @@ internal class UriToBitmapImageConverter : IValueConverter
 			try
 			{
 				var cacheKey = uri.AbsoluteUri;
-				if (ImageCache.TryGetValue(cacheKey, out var cachedBitmap))
+				lock (ImageCacheLock)
 				{
-					return cachedBitmap;
+					if (ImageCache.TryGetValue(cacheKey, out var cachedBitmap))
+					{
+						return cachedBitmap;
+					}
 				}
 
 				var bitmap = new BitmapImage();
+				void RemoveFailedImage(object _, System.Windows.Media.ExceptionEventArgs args)
+				{
+					lock (ImageCacheLock)
+					{
+						if (ImageCache.TryGetValue(cacheKey, out var failed) && ReferenceEquals(failed, bitmap))
+							ImageCache.Remove(cacheKey);
+					}
+					DivinityApp.Log($"Could not load a remote mod image from '{RedactRemoteImageUri(uri)}': {args.ErrorException?.Message}");
+				}
+				bitmap.DownloadFailed += RemoveFailedImage;
+				bitmap.DecodeFailed += RemoveFailedImage;
 				bitmap.BeginInit();
-				bitmap.CreateOptions = BitmapCreateOptions.PreservePixelFormat;
-				bitmap.CacheOption = BitmapCacheOption.OnDemand;
+				bitmap.CreateOptions = BitmapCreateOptions.PreservePixelFormat | BitmapCreateOptions.IgnoreImageCache;
+				bitmap.CacheOption = BitmapCacheOption.OnLoad;
 				bitmap.UriSource = uri;
 				bitmap.EndInit();
-				ImageCache[cacheKey] = bitmap;
+				lock (ImageCacheLock)
+				{
+					ImageCache[cacheKey] = bitmap;
+				}
 				return bitmap;
 			}
 			catch (Exception ex)
@@ -35,6 +53,12 @@ internal class UriToBitmapImageConverter : IValueConverter
 			}
 		}
 		return null;
+	}
+
+	private static string RedactRemoteImageUri(Uri uri)
+	{
+		if (uri == null || !uri.IsAbsoluteUri) return "remote image";
+		return $"{uri.Scheme}://{uri.Host}{uri.AbsolutePath}";
 	}
 
 	public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)